feat(document): 优化文档块查询接口实现
- 将TaskId字段标记为可选参数，支持按任务ID查询或使用文档当前任务
- 实现文档块分页查询功能，返回文档块列表、总数和计划ID
- 添加文档块详情查询功能，支持获取父块信息和兄弟块列表
- 新增DocumentChunkDetail聚合模型用于封装块详情数据结构
- 更新转换器实现，增加ToDocumentChunkItemList和ToQueryDocumentChunkDetailResp方法
- 重构实体模型，将枚举名称字段改为非数据库映射字段
- 优化任务查询逻辑，支持按任务类型筛选最新任务
- 重命名并整合任务相关枚举定义到统一的task_enum.go文件
- 调整数据访问层接口，支持更灵活的任务和块查询操作

+++ internal/infrastructure/model/document_chunk.go
@@ -8,17 +8,17 @@ type DocumentChunk struct {
 	TaskId            int64  `gorm:"column:task_id"`             // 任务ID
 	PlanId            int64  `gorm:"column:plan_id"`             // 计划ID
 	ParentBlockId     int64  `gorm:"column:parent_block_id"`     // 父块ID
-	ChunkNo           int64  `gorm:"column:chunk_no"`            // 块序号
-	SourceType        int64  `gorm:"column:source_type"`         // 来源类型
+	ChunkNo           int    `gorm:"column:chunk_no"`            // 块序号
+	SourceType        int    `gorm:"column:source_type"`         // 来源类型
 	SectionPath       string `gorm:"column:section_path"`        // 章节路径
 	StructureNodeId   int64  `gorm:"column:structure_node_id"`   // 结构节点ID
-	StructureNodeType int64  `gorm:"column:structure_node_type"` // 结构节点类型
+	StructureNodeType int    `gorm:"column:structure_node_type"` // 结构节点类型
 	CanonicalPath     string `gorm:"column:canonical_path"`      // 规范路径
 	ItemIndex         int64  `gorm:"column:item_index"`          // 项索引
 	ChunkText         string `gorm:"column:chunk_text"`          // 块文本
-	CharCount         int64  `gorm:"column:char_count"`          // 字符数
-	TokenCount        int64  `gorm:"column:token_count"`         // token数
-	VectorStatus      int64  `gorm:"column:vector_status"`       // 向量状态
-	VectorStoreType   int64  `gorm:"column:vector_store_type"`   // 向量存储类型
+	CharCount         int    `gorm:"column:char_count"`          // 字符数
+	TokenCount        int    `gorm:"column:token_count"`         // token数
+	VectorStatus      int    `gorm:"column:vector_status"`       // 向量状态
+	VectorStoreType   int    `gorm:"column:vector_store_type"`   // 向量存储类型
 	VectorId          string `gorm:"column:vector_id"`           // 向量ID
 }
